temptscrn: fix projector location for coap resources

+++ temp_tscrn/src/data_dispatcher.h
@@ -68,7 +68,7 @@ typedef struct {
         int16_t temp_measurement;
         int16_t temp_setting;
         uint16_t output;
-	uint16_t prj_validity;
+	uint32_t prj_validity;
 
         struct {
             data_ctlr_mode_t mode;
